requirements.txt
python-dotenv
groq
AppOpener
pywhatkit
bs4
pillow
rich
requests
keyboard
cohere
googlesearch-python
selenium
mtranslate
pygame
edge-tts
PyQt5
webdriver-manager
chromedriver-autoinstaller